+++ lastexpress.cpp
@@ -350,7 +350,7 @@ void LastExpressEngine::errorString(const char *buf_input, char *buf_output, int
 	snprintf(buf_output, buf_output_size, "%s", buf_input);
 }
 
-Scene *LastExpressEngine::getScene(uint16 sceneId) const {
+Scene *LastExpressEngine::getScene(uint32 sceneId) const {
 	return _sceneMan->getScene(sceneId);
 }
